Add missing files, premake fixes, shader fix
Noticed the missing files when building on my laptop, since it hasn't
been a problem on my desktop I guess they're not necessary for all
hardware configurations?

Premake5 updated to alpha12, added check for Windows SDK so it wont be
set to 8.1 which isn't installed by default on Windows 10.

Added location specifiers to passthrough.vert

Updated FlappyPong scene in Examples to remove Javascript stuff
(generally I've been working on a FlappyPong project outside the tree
which has made this one stagnate a bit)