Change include structure
Now basically works OOTB with vcpkg includes instead of needing to manually
copy to the include directory. Only things missing are moodycamel queue and
Vulkan headers. Those still need to be manually downloaded and added.

There is also an issue in vcpkg's Bullet packaging where the include files
are in a bullet/ subdirectory of the include folder, but the bullet files
assume they are at the root of the include directory. To fix this you
need to copy the files from include/bullet into include/.