Initial UI and editor support
The Imgui render function is pretty poorly optimized right now but it
does the job.

The editor is opened by pressing F7.
The editor allows editing component properties and dragging the camera
around by holding ALT+Left click while dragging the mouse. The camera
position is reset when the editor UI is hidden.

There is a known issue with OpenGL rendering of the GUI that causes the
Nvidia driver to crash... So OpenGL is borked for now.