Refactor deserialization
Deserializable types can now provide a schema that the SerializedObject
will be checked against before the Deserializer is called. This allows
the Deserializer to focus on reading values instead of also having to
check for schema violations, as well as providing uniform error
messages for all violations. This schema will also form the backbone
for more features in the editor.

The schema currently lacks support for referring to other schemas.

Also made Scene Deserializable to take advantage of the schema
validator.