Remove SubmitMesh
SubmitMesh resulted in a lot of copying in StaticMeshComponent when the
array of submeshes was copied. Now, only a small object with updates
needs to be copied each frame, and it should be possible to remove that
as well.

This also has the side effect of making it easy to toggle rendering of
a specific StaticMeshComponent on and off in the editor.