Fix errors when hot reloading sprites
After refactoring sprite logic to reside in RenderSystem, the hot reload
code did not keep up, so there would be errors caused by using an out of
date ImageView object in a DescriptorSet. This is fixed by making it so
the SpriteComponent notifies the RenderSystem when the Image has been
changed.