Massive refactor of Scene/GameModule
My original thinking here was "Scene should not contain input config"
and then it grew into quite a bit more than that.

The primary change in this commit is to reduce the responsibilities of
the Scene class and the GameModule namespace. Previously it was unclear
what these two things were and weren't responsible for.

The new thinking is that configuration that is global for a game, such
as keybindings, which DLLs to load, etc. no longer goes in the Scene
object. That type of configuration goes in a new class called Project.
The Project class and the updated Scene class are fairly "dumb" classes
that mostly just translate JSON files to C++ objects.

The GameModule is no longer responsible for keeping track of the
PhysicsWorld or loading DLLs. PhysicsWorld is now a singleton with a
GetInstance method and DLLs are loaded using a new DllManager singleton.

This commit also changes how entities are allocated, they are now
handled by a new EntityManager singleton. Instead of having raw Entity
pointers you now use an EntityPtr object to reference Entities. The
EntityPtr currently doesn't do much, but in the future it may become
more of a smart pointer type thing.

This commit also:
* Upgrades the project to "c++latest" and starts using named initializer
* Fixes a crash when deleting SkeletalMeshComponent
* Disables buttons in the editor when the editor is in a state where
they can't be used (for example, disables the Save Scene button if no
Scene is loaded)
* Adds reflection support for std::string and raw pointers
* Adds an id property to all entities which can be used to uniquely
identify entities even if they have the same name.