Fix alt-drag in editor not caring abour rotation
Previously alt-drag would always move the camera along worldspace x, y.
For some reason it has been difficult to get this to behave properly and
even in this commit a safety net of checking for nan/inf before using
the position/rotation is used. For some reason alt-dragging sometimes
causes the position to become full of NaNs, which is now ignored.