Rewrite Logger to use <format>
Also removes the use of std::unique_ptr in Logger.

This was a pain in the ass since every place that uses Logger had to be
updated. Hopefully similar changes to Logger will not be needed for a
while.

The benefits of this change are:
* Remove the extra indirection of all loggers being pointers
* Remove the need to call .c_str on strings/paths when logging them

The downside of this change is:
*  The implementation of logger is no longer hidden in the .cpp file.
   This means that any change to the implementation will basically lead
   to a recompile of the whole engine...

I consider this a worthwhile tradeoff because changes to Logger are rare
but calls to the logger with types like string and optional are fairly
common, and this change reduces how much code is needed for all calls.

Logging that previously had to look like
`logger->Infof("myString %s", myString.c_str())` now looks like
`logger.Info("myString {}", myString)`
and logging that looked like
`logger->Infof("myOpt %s", (myOpt.has_value() ? myOpt.value().c_str() : "")`
now looks like `logger.Info("myOpt {}", myOpt)`.