Handle shaderc like other third party dependencies
Previously shaderc needed to be installed through a package manager like
vcpkg. Now it is instead checked out as a submodule and built along with
the engine.

Unfortunately these dependencies require Python in order to build which
is why I was hesitant to add them. But now I believe that the tradeoff
is worth it.